# openapi/contabilidad-api.yaml
openapi: 3.1.0
info:
  title: Contabilidad API
  description: |
    API REST para el sistema de contabilidad y facturación electrónica.
    Soporta CFDI 4.0, REP 2.0, catálogos SAT versionados, motor fiscal configurable
    y preparación para contabilidad electrónica.
  version: 0.1.0
  contact:
    name: Contabilidad Team
servers:
  - url: http://localhost:8080
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/clients:
    get:
      tags: [Clients]
      summary: Listar clientes
      operationId: listClients
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
        - name: search
          in: query
          schema:
            type: string
          description: Búsqueda por RFC o nombre
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientPage'
    post:
      tags: [Clients]
      summary: Crear cliente
      operationId: createClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
      responses:
        '201':
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDto'
        '409':
          description: RFC duplicado en la empresa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '422':
          description: Validación fiscal fallida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/clients/{clientId}:
    get:
      tags: [Clients]
      summary: Obtener cliente
      operationId: getClient
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cliente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDto'
    patch:
      tags: [Clients]
      summary: Actualizar cliente
      operationId: updateClient
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClientRequest'
      responses:
        '200':
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDto'
    delete:
      tags: [Clients]
      summary: Eliminar cliente (soft delete)
      operationId: deleteClient
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Cliente eliminado
...
  /api/v1/suppliers:
    get:
      tags: [Suppliers]
      summary: Listar proveedores
      operationId: listSuppliers
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: Lista de proveedores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierPage'
    post:
      tags: [Suppliers]
      summary: Crear proveedor
      operationId: createSupplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupplierRequest'
      responses:
        '201':
          description: Proveedor creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierDto'
        '409':
          description: RFC duplicado
        '422':
          description: Validación fallida
  /api/v1/suppliers/{supplierId}:
    get:
      tags: [Suppliers]
      summary: Obtener proveedor
      operationId: getSupplier
      parameters:
        - name: supplierId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Proveedor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierDto'
    patch:
      tags: [Suppliers]
      summary: Actualizar proveedor
      operationId: updateSupplier
      parameters:
        - name: supplierId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSupplierRequest'
      responses:
        '200':
          description: Proveedor actualizado
    delete:
      tags: [Suppliers]
      summary: Eliminar proveedor
      operationId: deleteSupplier
      parameters:
        - name: supplierId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Proveedor eliminado

  # ── Products ──
  /api/v1/products:
    get:
      tags: [Products]
      summary: Listar productos
      operationId: listProducts
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: Lista de productos
    post:
      tags: [Products]
      summary: Crear producto
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Producto creado
  /api/v1/products/{productId}:
    get:
      tags: [Products]
      summary: Obtener producto
      operationId: getProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Producto
    patch:
      tags: [Products]
      summary: Actualizar producto
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Producto actualizado
    delete:
      tags: [Products]
      summary: Eliminar producto
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Producto eliminado

  # ── Catalogs ──
...
  /api/v1/payments:
    get:
      tags: [Payments]
      summary: Listar pagos
      operationId: listPayments
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Lista de pagos
    post:
      tags: [Payments]
      summary: Registrar pago
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Pago registrado
  /api/v1/payments/{paymentId}:
    get:
      tags: [Payments]
      summary: Obtener pago
      operationId: getPayment
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Pago
  /api/v1/payments/{paymentId}/apply:
    post:
      tags: [Payments]
      summary: Aplicar pago contra documento
      operationId: applyPayment
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyPaymentRequest'
      responses:
        '200':
          description: Pago aplicado

  # ── Reports ──
  /api/v1/reports/income-expense:
    get:
      tags: [Reports]
      summary: Reporte de ingresos y egresos
      operationId: getIncomeExpenseReport
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Reporte
  /api/v1/reports/taxes:
    get:
      tags: [Reports]
      summary: Reporte de impuestos
      operationId: getTaxReport
      parameters:
        - name: periodKey
          in: query
          required: true
          schema:
            type: string
            description: "Formato: YYYY-MM"
      responses:
        '200':
          description: Reporte de impuestos
  /api/v1/reports/aging:
    get:
      tags: [Reports]
      summary: Reporte de antigüedad de saldos
      operationId: getAgingReport
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [AR, AP]
      responses:
        '200':
          description: Reporte de antigüedad
  /api/v1/reports/client-summary:
    get:
      tags: [Reports]
      summary: Resumen por cliente
      operationId: getClientSummary
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Resumen por cliente
  /api/v1/reports/supplier-summary:
    get:
      tags: [Reports]
      summary: Resumen por proveedor
      operationId: getSupplierSummary
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Resumen por proveedor

  # ── Declarations ──
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    size:
      name: size
      in: query
      schema:
        type: integer
        default: 10
    sort:
      name: sort
      in: query
      schema:
        type: string
        description: "Campo y dirección, ej: legalName,asc"
  schemas:
...
    ClientDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        rfc:
          type: string
        legalName:
          type: string
        tradeName:
          type: string
        email:
          type: string
        phone:
          type: string
        website:
          type: string
        fiscalRegimeCode:
          type: string
        defaultUsoCfdiCode:
          type: string
        defaultFormaPagoCode:
          type: string
        defaultMetodoPagoCode:
          type: string
        defaultPostalCode:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ClientPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ClientDto'
        page:
          type: integer
        size:
          type: integer
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
    CreateClientRequest:
      type: object
      required: [rfc, legalName, fiscalRegimeCode]
      properties:
        rfc:
          type: string
        legalName:
          type: string
        email:
          type: string
        phone:
          type: string
        website:
          type: string
        fiscalRegimeCode:
          type: string
        defaultUsoCfdiCode:
          type: string
        defaultFormaPagoCode:
          type: string
        defaultPostalCode:
          type: string
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/AddressRequest'
    UpdateClientRequest:
      type: object
      properties:
        legalName:
          type: string
        email:
          type: string
        phone:
          type: string
        website:
          type: string
        fiscalRegimeCode:
          type: string
        defaultUsoCfdiCode:
          type: string
        defaultFormaPagoCode:
          type: string
        defaultPostalCode:
          type: string
    AddressRequest:
      type: object
      required: [street1, postalCode]
      properties:
        street1:
          type: string
        street2:
          type: string
        exteriorNumber:
          type: string
        interiorNumber:
          type: string
        neighborhood:
          type: string
        city:
          type: string
        municipalityCode:
          type: string
        stateCode:
          type: string
        postalCode:
          type: string
        countryCode:
          type: string
          default: MEX

    # ── Suppliers ──
    SupplierDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        rfc:
          type: string
        legalName:
          type: string
        email:
          type: string
        phone:
          type: string
        fiscalRegimeCode:
          type: string
        nationality:
          type: string
        active:
          type: boolean
    SupplierPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SupplierDto'
        page:
          type: integer
        size:
          type: integer
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
    CreateSupplierRequest:
      type: object
      required: [rfc, legalName, fiscalRegimeCode]
      properties:
        rfc:
          type: string
        legalName:
          type: string
        email:
          type: string
        phone:
          type: string
        fiscalRegimeCode:
          type: string
        nationality:
          type: string
          default: NATIONAL

    # ── Products ──
    CreateProductRequest:
      type: object
      required: [internalName, satProductCode, satUnitCode, unitPrice]
      properties:
        internalName:
          type: string
        description:
          type: string
        satProductCode:
          type: string
        satUnitCode:
          type: string
        unitPrice:
          type: number
          format: decimal
        currencyCode:
          type: string
          default: MXN
        objetoImpCode:
          type: string
          default: '02'
        taxProfiles:
          type: array
          items:
            $ref: '#/components/schemas/TaxProfileRequest'
    TaxProfileRequest:
      type: object
      required: [taxCode, factorType, rate, validFrom]
      properties:
        taxCode:
          type: string
        factorType:
          type: string
        rate:
          type: number
          format: decimal
        isTransfer:
          type: boolean
          default: true
        isWithholding:
          type: boolean
          default: false
        validFrom:
          type: string
          format: date

    # ── Invoices ──
...
    CreatePaymentRequest:
      type: object
      required: [paymentDirection, paymentFormCode, amount, paidAt]
      properties:
        paymentDirection:
          type: string
          enum: [INBOUND, OUTBOUND]
        paymentFormCode:
          type: string
        currencyCode:
          type: string
          default: MXN
        amount:
          type: number
        paidAt:
          type: string
          format: date-time
        operationNumber:
          type: string
        notes:
          type: string
    ApplyPaymentRequest:
      type: object
      required: [documentType, documentId, amountPaid]
      properties:
        documentType:
          type: string
          enum: [INVOICE, EXPENSE]
        documentId:
          type: string
          format: uuid
        amountPaid:
          type: number

    # ── Declarations ──
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              rejectedValue: {}